Exceptions delete e update

+++ src/main/java/com/lsmartin/crudcliente/services/ClientService.java
@@ -2,6 +2,7 @@ package com.lsmartin.crudcliente.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.lsmartin.crudcliente.dto.ClientDTO;
@@ -56,6 +57,11 @@ public class ClientService {
 	
 	@Transactional
 	public  void delete(Long id) {	
+		
+		if(!clientRepository.existsById(id)) {
+			
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 		
 		clientRepository.deleteById(id);;
